backend/requirements.txt
Django>=4.2
psycopg2-binary>=2.9
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
django-cors-headers>=4.0 
gunicorn>=20.1.0
Pillow==11.2.1
argon2-cffi>=21.3.0
